refactor(api): PR-6 commit 2 — 4 admin router (jobs/secrets/models/upscalers)
server.py 3944 → 3786 行（−158），10 routes + 3 inline BaseModel 抽出 +
新增 api/deps.py 给跨 router 共享 helper。

  api/routers/jobs.py        3 routes  GET /api/jobs/{jid}, /log; POST /cancel
  api/routers/secrets.py     2 routes  GET / PUT /api/secrets
  api/routers/models.py      3 routes  /api/models/{catalog,path-defaults,download}
  api/routers/upscalers.py   2 routes  /api/upscalers/{select,download_custom}

  api/schemas/models.py     3 BaseModel  ModelDownloadRequest /
                                        UpscalerSelectRequest /
                                        UpscalerCustomDownloadRequest

api/deps.py 新建（Agent A 原方案有这文件，PR-5 暂未建；PR-6 commit 2
建出来给 jobs router 用 _supervisor()）：
  _supervisor()  从 app.state 取 Supervisor（late import 避开 api.app ↔
                 api.routers.* ↔ api.deps 三方循环）

server.py 内：
  - 删 `_supervisor()` 函数体，改 `from .api.deps import _supervisor`
    保剩余 queue / generate / projects router 内 callsite 不动（PR-6.5 之前
    都还在 server.py）
  - secrets router 用 `from .. import secrets`，避免跟 `secrets_router`
    名字撞（app.py include 时 `from .routers import secrets as secrets_router`）

回归：PR-1 安全网 152 + studio_server 44 + supervisor_jobs 7 +
model_downloader 27 + secrets 2 + ... = **289/289 全绿**。route snapshot 不变
（160 routes）。
